requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
confusable-homoglyphs==3.1.1
decorator==4.3.0
Django==2.0.5
django-multiselectfield==0.1.8
django-registration==2.4.1
django-sass-processor==0.7
djangorestframework==3.8.2
factory-boy==2.11.1
Faker==0.8.15
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
jsonschema==2.6.0
jupyter-core==4.4.0
lxml==4.2.1
nbformat==4.4.0
numpy==1.14.3
pandas==0.23.0
pandas-datareader==0.6.0
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
plotly==2.6.0
prompt-toolkit==1.0.15
psycopg2==2.7.4
psycopg2-binary==2.7.4
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
requests-file==1.4.3
requests-ftp==0.3.1
simplegeneric==0.8.1
six==1.11.0
text-unidecode==1.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
